+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Workoutizer 0.15.0 (2021-03-21)
 
 ### Features
-* Parse best climb sections from fit files using [sportgems](https://github.com/fgebhart/sportgems) and integrate it
-  into awards and activity page.
+* Parse best climb sections from activity files using [sportgems](https://github.com/fgebhart/sportgems) and integrate it
+  into awards and activity page (next to fastest sections).
 
 
 ## Workoutizer 0.14.0 (2021-03-15)